requirements.txt
asgiref==3.7.1
async-timeout==4.0.2
attrs==23.1.0
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
django-redis==5.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
inflection==0.5.1
jsonschema==4.17.3
psycopg2-binary==2.9.6
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.5
sqlparse==0.4.4
typing_extensions==4.6.2
uritemplate==4.1.1